client/requirements.txt
netconf
pyasn1
pyasn1-modules
cryptography
pyopenssl
certvalidator
asn1crypto
